feat: TOOLS-2560 add batch deletes

+++ src/main/transaction.c
@@ -105,7 +105,13 @@ LOCAL_HELPER as_batch_records*
 _gen_batch_writes_random_keys(const cdata_t* cdata, tdata_t* tdata,	
 		const stage_t* stage);
 LOCAL_HELPER void throttle(tdata_t* tdata, thr_coord_t* coord);
-
+LOCAL_HELPER as_batch_records* _gen_batch_deletes_random_keys(
+		const cdata_t* cdata, tdata_t* tdata, const stage_t* stage);
+LOCAL_HELPER as_batch_records* _gen_batch_deletes_sequential_keys(
+		const cdata_t* cdata, tdata_t* tdata, const stage_t* stage, uint64_t start_key);
+LOCAL_HELPER as_batch_records* _gen_batch_deletes(
+		const cdata_t* cdata, tdata_t* tdata, const stage_t* stage, bool randomKeys,
+		uint64_t start_key);
 // Synchronous workload helper methods
 LOCAL_HELPER void random_read(tdata_t* tdata, cdata_t* cdata,
 		thr_coord_t* coord, const stage_t* stage);
@@ -762,6 +768,68 @@ _gen_batch_writes(const cdata_t* cdata, tdata_t* tdata,
 }
 
 /*
+ * generates a batch of keys for use with aerospike_batch_remove
+ * keys are generated randomly between stage->key_start ad stage->key_end
+ */
+LOCAL_HELPER inline as_batch_records*
+_gen_batch_deletes_random_keys(const cdata_t* cdata, tdata_t* tdata,	
+		const stage_t* stage)
+{
+	return _gen_batch_deletes(cdata, tdata, stage, true, stage->key_start);
+}
+
+/*
+ * generates a batch of keys for use with aerospike_batch_remove
+ * keys used in the batch remove will be sequential from key_start
+ */
+LOCAL_HELPER inline as_batch_records*
+_gen_batch_deletes_sequential_keys(const cdata_t* cdata, tdata_t* tdata,	
+		const stage_t* stage, uint64_t start_key)
+{
+	return _gen_batch_deletes(cdata, tdata, stage, false, start_key);
+}
+
+/*
+ * generates a batch for use with aerospike_batch_remove.
+ * if randomKeys == false the keys used in the batch will be sequential from key_start
+ * otherwise keys are generated randomly between stage->key_start and stage->key_end
+ * this function should only be called through its wrappers _gen_batch_deletes_random_keys
+ * and _gen_batch_deletes_sequential_keys
+ */
+LOCAL_HELPER as_batch_records*
+_gen_batch_deletes(const cdata_t* cdata, tdata_t* tdata,	
+		const stage_t* stage, bool randomKeys, uint64_t start_key)
+{
+	uint32_t batch_size = stage->batch_delete_size;
+	uint64_t key_val = start_key;
+
+	as_batch_records* batch = as_batch_records_create(batch_size);
+
+	for (uint32_t i = 0; i < batch_size; i++) {
+		as_record* rec;
+		rec = _gen_nil_record(tdata);
+		as_batch_remove_record* batch_remove = as_batch_remove_reserve(batch);
+		// set the batch remove key value pointer to the address of its own
+		// value so that when the key is initialized, its value is stored
+		// in batch->key.value
+		batch_remove->key.valuep = &batch_remove->key.value;
+
+		if (randomKeys) {
+			key_val = stage_gen_random_key(stage, tdata->random);
+			_gen_key(key_val, &batch_remove->key, cdata);
+		}
+		else {
+			_gen_key(key_val, &batch_remove->key, cdata);
+			++key_val;
+		}
+
+		batch_remove->policy = &tdata->policies.batch_remove;
+	}
+
+	return batch;
+}
+
+/*
  * generates a record with all nil bins (used to remove records)
  */
 LOCAL_HELPER as_record*
@@ -893,21 +961,36 @@ LOCAL_HELPER void
 random_delete(tdata_t* tdata, cdata_t* cdata, thr_coord_t* coord,
 		const stage_t* stage)
 {
-	as_key key;
-	as_record* rec;
 
-	// generate a random key
-	uint64_t key_val = stage_gen_random_key(stage, tdata->random);
-	_gen_key(key_val, &key, cdata);
+	if (stage->batch_delete_size >= 1) {
+		as_key key;
+		as_record* rec;
 
-	// create a record
-	rec = _gen_nil_record(tdata);
+		// generate a random key
+		uint64_t key_val = stage_gen_random_key(stage, tdata->random);
+		_gen_key(key_val, &key, cdata);
 
-	// write this record to the database
-	_write_record_sync(tdata, cdata, coord, &key, rec);
+		// create a record
+		rec = _gen_nil_record(tdata);
 
-	// don't destroy delete records
-	as_key_destroy(&key);
+		// write this record to the database
+		_write_record_sync(tdata, cdata, coord, &key, rec);
+
+		// don't destroy delete records
+		as_key_destroy(&key);
+	}
+	else {
+		as_batch_records* batch;
+
+		batch = _gen_batch_deletes_random_keys(cdata, tdata, stage);
+		_batch_write_record_sync(tdata, cdata, coord, batch);
+
+		for (uint32_t i = 0; i < batch->list.size; i++) {
+			as_batch_write_record* r = as_vector_get(&batch->list, i);
+			as_operations_destroy(r->ops);
+		}
+		as_batch_records_destroy(batch);
+	}
 }
 
 
@@ -1047,17 +1130,32 @@ linear_deletes(tdata_t* tdata, cdata_t* cdata, thr_coord_t* coord,
 	while (tdata->do_work &&
 			key_val < end_key) {
 
-		// create a record with given key
-		_gen_key(key_val, &key, cdata);
-		rec = _gen_nil_record(tdata);
+		if (stage->batch_delete_size <= 1) {
+			// create a record with given key
+			_gen_key(key_val, &key, cdata);
+			rec = _gen_nil_record(tdata);
 
-		// delete this record from the database
-		_write_record_sync(tdata, cdata, coord, &key, rec);
+			// delete this record from the database
+			_write_record_sync(tdata, cdata, coord, &key, rec);
 
-		_destroy_record(rec, stage);
-		as_key_destroy(&key);
+			_destroy_record(rec, stage);
+			as_key_destroy(&key);
 
-		key_val++;
+			key_val++;
+		}
+		else {
+			as_batch_records* batch;
+
+			batch = _gen_batch_deletes_sequential_keys(cdata, tdata, stage, key_val);
+			_batch_write_record_sync(tdata, cdata, coord, batch);
+			key_val += stage->batch_delete_size;  // TODO change to batch size
+
+			for (uint32_t i = 0; i < batch->list.size; i++) {
+				as_batch_write_record* r = as_vector_get(&batch->list, i);
+				as_operations_destroy(r->ops);
+			}
+			as_batch_records_destroy(batch);
+		}
 	}
 
 	// once we've written everything, there's nothing left to do, so tell
@@ -1179,18 +1277,32 @@ LOCAL_HELPER void
 random_delete_async(tdata_t* tdata, cdata_t* cdata, thr_coord_t* coord,
 		const stage_t* stage, struct async_data_s* adata)
 {
-	as_record* rec;
+	if (stage->batch_delete_size <= 1) {
+		as_record* rec;
 
-	// generate a random key
-	uint64_t key_val = stage_gen_random_key(stage, tdata->random);
+		// generate a random key
+		uint64_t key_val = stage_gen_random_key(stage, tdata->random);
 
-	_gen_key(key_val, &adata->key, cdata);
-	rec = _gen_nil_record(tdata);
-	adata->op = write_op;
+		_gen_key(key_val, &adata->key, cdata);
+		rec = _gen_nil_record(tdata);
+		adata->op = write_op;
 
-	_write_record_async(&adata->key, rec, adata, tdata, cdata);
+		_write_record_async(&adata->key, rec, adata, tdata, cdata);
 
-	_destroy_record(rec, stage);
+		_destroy_record(rec, stage);
+	}
+	else {
+		as_batch_records* batch;
+
+		batch = _gen_batch_deletes_random_keys(cdata, tdata, stage);
+		_batch_write_record_async(tdata, cdata, coord, batch);
+
+		for (uint32_t i = 0; i < batch->list.size; i++) {
+			as_batch_write_record* r = as_vector_get(&batch->list, i);
+			as_operations_destroy(r->ops);
+		}
+		as_batch_records_destroy(batch);
+	}
 }
 
 
@@ -1494,27 +1606,42 @@ linear_deletes_async(tdata_t* tdata, cdata_t* cdata, thr_coord_t* coord,
 	while (tdata->do_work &&
 			key_val < end_key) {
 
-		adata = queue_pop_wait(adata_q);
+		if (stage->batch_delete_size <= 1) {
+			adata = queue_pop_wait(adata_q);
 
-		clock_gettime(COORD_CLOCK, &wake_time);
-		start_time = timespec_to_us(&wake_time);
-		adata->start_time = start_time;
+			clock_gettime(COORD_CLOCK, &wake_time);
+			start_time = timespec_to_us(&wake_time);
+			adata->start_time = start_time;
 
-		as_record* rec;
-		_gen_key(key_val, &adata->key, cdata);
-		rec = _gen_nil_record(tdata);
-		adata->op = delete_op;
+			as_record* rec;
+			_gen_key(key_val, &adata->key, cdata);
+			rec = _gen_nil_record(tdata);
+			adata->op = delete_op;
 
-		_write_record_async(&adata->key, rec, adata, tdata, cdata);
+			_write_record_async(&adata->key, rec, adata, tdata, cdata);
 
-		_destroy_record(rec, stage);
+			_destroy_record(rec, stage);
 
-		uint64_t pause_for =
-			dyn_throttle_pause_for(&tdata->dyn_throttle, start_time);
-		timespec_add_us(&wake_time, pause_for);
-		thr_coordinator_sleep(coord, &wake_time);
+			uint64_t pause_for =
+				dyn_throttle_pause_for(&tdata->dyn_throttle, start_time);
+			timespec_add_us(&wake_time, pause_for);
+			thr_coordinator_sleep(coord, &wake_time);
 
-		key_val++;
+			key_val++;
+		}
+		else {
+			as_batch_records* batch;
+
+			batch = _gen_batch_deletes_sequential_keys(cdata, tdata, stage, key_val);
+			_batch_write_record_async(batch, adata, tdata, cdata);
+			key_val += stage->batch_delete_size;
+
+			for (uint32_t i = 0; i < batch->list.size; i++) {
+				as_batch_write_record* r = as_vector_get(&batch->list, i);
+				as_operations_destroy(r->ops);
+			}
+			as_batch_records_destroy(batch);
+		}
 	}
 
 	// once we've written everything, there's nothing left to do, so tell
